Moved the API download to the Show Activity and changed it to description.
