+++ app/src/main/res/layout/activity_show.xml
@@ -45,7 +45,7 @@
 
     <TextView
         android:id="@+id/tvShowDescription"
-        android:layout_width="wrap_content"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
         android:layout_marginLeft="8dp"
@@ -53,7 +53,6 @@
         android:layout_marginEnd="8dp"
         android:layout_marginRight="8dp"
         android:gravity="center"
-        android:text="@string/show_description"
         android:textColor="@android:color/white"
         android:textSize="18sp"
         app:fontFamily="@font/lato"
